Renamed the Scenes folder to _Scenes.
Adding the underscore to the scenes folder will bring it nearly to the top of the Assets folder. There are a few things that will come before it, like assets with names that start with a digit. Yet, even then, doing this to the scenes folder is quite handy as the number of items directly in the Assets folder goes up.