One of the files in .git got currupted and I had to checkout the copy in GitHub to contiune working.
Since I checked the copy on the GitHub server, some things got changed during checkout. Some of them are as simple as GitHub didn't think to put in an exception for their down pdb files so only their meta files were checked in and are now being removed by Unity since there are no pdb filles said meta files. Other items, The newline character was changed, I'm guessing either the GitHub server, or the GitHub plugin, or my Git client did that, so 2 of 3 odds says that it was GitHubs fault. So far, their plugin has created more issues then it has fixed. Really, the only thing it has going for it is the auto LFS configuration and their rule setup for that still needs some work.