Getting rid of GitHub plugin.
Why? Because, it had me update the plugin, and updating it created a list of changes that it refuses to commit. In fact, this plugin has a tendency to refuse to commit changes. It has a tendency to refuse to acknowledge changes. It has a tendency to break down when there are too many changes. It provides poor feedback when something fails to commit. If provides poor feedback when something fails to push.